# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Crypto Trade API
  version: 1.0.0
  description: Sandbox crypto trading API for students. Covers market data, wallets, buy, sell, swap, transactions, KYC, and admin operations.
servers:
  - url: /
    description: Current API host
  - url: http://localhost:4200
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /admin/users/{userId}:
    get:
      tags: [Admin]
      summary: Get customer detail, wallet, and transactions
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/AdminUserDetailResponse"
...
  /admin/kyc/{kycId}:
    patch:
      tags: [Admin]
      summary: Approve or reject KYC
      parameters:
        - in: path
          name: kycId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, rejected]
                  example: approved
                reviewerNote:
                  type: string
                  example: Approved for classroom demo.
      responses:
        "200":
          description: Reviewed KYC
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/KycResponse"
...
  /admin/withdrawals/{withdrawalId}:
    patch:
      tags: [Admin]
      summary: Approve or reject a withdrawal
      parameters:
        - in: path
          name: withdrawalId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, rejected]
                  example: approved
                reviewerNote:
                  type: string
                  example: Approved after review.
      responses:
        "200":
          description: Reviewed withdrawal
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/WithdrawalResponse"
  /admin/assets:
    get:
      tags: [Admin]
      summary: List managed assets
      responses:
        "200":
          description: Assets
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/AssetListResponse"
    post:
      tags: [Admin]
      summary: Add a managed asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol, name, network, priceUsd]
              properties:
                symbol:
                  type: string
                  example: SOL
                name:
                  type: string
                  example: Solana
                network:
                  type: string
                  example: Solana Devnet
                priceUsd:
                  type: number
                  example: 142.25
                change24h:
                  type: number
                  example: 2.4
                isActive:
                  type: boolean
                  example: true
                minBuyUsd:
                  type: number
                  example: 10
                minSellUsd:
                  type: number
                  example: 10
                iconUrl:
                  type: string
                  example: /assets/sol.svg
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/AssetResponse"
  /admin/assets/{symbol}:
    patch:
      tags: [Admin]
      summary: Pause or allow an asset
      parameters:
        - $ref: "#/components/parameters/Symbol"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive:
                  type: boolean
                  example: false
      responses:
        "200":
          description: Updated asset availability
          content:
            application/json:
              examples:
                paused:
                  $ref: "#/components/examples/AssetPausedResponse"
                allowed:
                  $ref: "#/components/examples/AssetResponse"
  /admin/fees:
    get:
      tags: [Admin]
      summary: Get trading fee settings
      responses:
        "200":
          description: Fee settings
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/FeeSettingsResponse"
    patch:
      tags: [Admin]
      summary: Update trading fee settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                buyFeePercent:
                  type: number
                  example: 1
                sellFeePercent:
                  type: number
                  example: 1.2
                swapFeePercent:
                  type: number
                  example: 0.8
                withdrawalFlatUsd:
                  type: number
                  example: 2
                spreadPercent:
                  type: number
                  example: 0.4
      responses:
        "200":
          description: Fee settings
          content:
            application/json:
              examples:
                success:
                  $ref: "#/components/examples/FeeSettingsResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Symbol:
      in: path
      name: symbol
      required: true
      schema:
        type: string
      example: BTC
  examples:
...
    KycResponse:
      summary: KYC submission
      value:
        data:
          id: kyc_student
          userId: usr_student
          legalName: Ada Student
          country: Nigeria
          documentType: national_id
          documentNumber: NIN-000-000
          selfieImageUrl: https://images.unsplash.com/photo-1494790108377-be9c29b29330?w=240&h=240&fit=crop&crop=faces
          documentImageUrl: https://images.unsplash.com/photo-1586953208448-b95a79798f07?w=480&h=300&fit=crop
          status: pending
          submittedAt: "2026-05-03T14:05:00.000Z"
          reviewedAt: null
          reviewerNote: null
    AssetResponse:
      summary: Crypto asset
      value:
        data:
          id: asset_btc
          symbol: BTC
          name: Bitcoin
          network: Bitcoin Testnet
          priceUsd: 64200.5
          change24h: 2.1
          isActive: true
          minBuyUsd: 10
          minSellUsd: 10
          iconUrl: /assets/btc.svg
    AssetPausedResponse:
      summary: Paused asset
      value:
        data:
          id: asset_btc
          symbol: BTC
          name: Bitcoin
          network: Bitcoin Testnet
          priceUsd: 64200.5
          change24h: 2.1
          isActive: false
          minBuyUsd: 10
          minSellUsd: 10
          iconUrl: /assets/btc.svg
    AssetListResponse:
      summary: Asset list
      value:
        data:
          - id: asset_btc
            symbol: BTC
            name: Bitcoin
            network: Bitcoin Testnet
            priceUsd: 64200.5
            change24h: 2.1
            isActive: true
            minBuyUsd: 10
            minSellUsd: 10
            iconUrl: /assets/btc.svg
          - id: asset_eth
            symbol: ETH
            name: Ethereum
            network: Ethereum Sepolia
            priceUsd: 3100.25
            change24h: -0.8
            isActive: true
            minBuyUsd: 10
            minSellUsd: 10
            iconUrl: /assets/eth.svg
        meta:
          count: 2
...
    WithdrawalResponse:
      summary: Withdrawal request
      value:
        data:
          id: wd_001
          userId: usr_student
          assetSymbol: ETH
          amount: 0.05
          feeAssetAmount: 0.00064511
          address: "0xabc123abc123abc123abc123abc123abc123abcd"
          network: Ethereum Sepolia
          status: pending
          createdAt: "2026-05-03T14:20:00.000Z"
          reviewedAt: null
          reviewerNote: null
...
    AdminUserDetailResponse:
      summary: Customer detail
      value:
        data:
          user:
            id: usr_student
            role: customer
            fullName: Ada Student
            email: student@cryptoclass.test
            phone: "+2348010000001"
            twoFactorEnabled: false
            kycStatus: approved
            avatarUrl: null
            watchlist: [BTC, ETH]
            settings:
              language: en
              fiatCurrency: USD
              theme: system
              priceAlerts: true
              pushNotifications: true
              biometricEnabled: false
            createdAt: "2026-05-03T14:00:00.000Z"
          wallet:
            id: wallet_student
            userId: usr_student
            fiatCurrency: USD
            depositAddresses: []
            balances:
              - assetSymbol: USD
                available: 1000
                locked: 0
          portfolioValueUsd: 1000
          transactions: []
...
    FeeSettingsResponse:
      summary: Fee settings
      value:
        data:
          buyFeePercent: 1
          sellFeePercent: 1.2
          swapFeePercent: 0.8
          withdrawalFlatUsd: 2
          spreadPercent: 0.4